requirements.txt
Django==2.0
python-decouple==3.1
pytz==2017.3
fabric3
